RevoBank-api/requirements.txt
flask==2.1.0
flask-sqlalchemy==2.5.1
flask-marshmallow==0.14.0
flask-bcrypt==1.0.1
flask-jwt-extended==4.4.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
marshmallow-sqlalchemy==0.28.0
email-validator==1.2.1